requirements.txt
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==3.0
pypdfium2==5.3.0
PyYAML==6.0.3
regex==2026.1.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
